requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.9.1.post1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.13
faiss-cpu==1.7.4
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.238
langsmith==0.0.11
loguru==0.7.0
marshmallow==3.19.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pyaes==1.6.1
pydantic==1.10.11
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
tenacity==8.2.2
TgCrypto==1.2.5
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.4
uvloop==0.17.0
yarl==1.9.2
